requirements.txt
absl-py==2.3.1
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
astunparse==1.6.3
attrs==24.3.0
av==14.0.1
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.2.0
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
chromadb==1.1.0
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cssselect==1.4.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
dgl==2.2.1
distlib==0.3.9
distro==1.9.0
durationpy==0.10
einops==0.8.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.118.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.16.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.7
kubernetes==34.1.0
langchain==0.3.27
langchain-anthropic==0.3.17
langchain-community==0.3.30
langchain-core==0.3.77
langchain-google-genai==2.0.10
langchain-huggingface==0.3.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.11
langcodes==3.5.0
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.8
language_data==1.3.0
libclang==18.1.1
lmdb==1.5.1
lxml==6.0.2
lxml_html_clean==0.4.3
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.1
mmh3==5.2.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.48.0
networkx==3.3
newspaper3k==0.2.8
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.0
open_clip_torch==2.30.0
openai==1.97.0
opencv-python==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.0
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.8
posthog==5.4.0
preshed==3.0.9
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.4
pypdf==6.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.9.11
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.2
scikeras==0.13.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==75.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stanza==1.10.1
starlette==0.48.0
streamlit==1.47.0
sympy==1.13.1
tbb==2021.11.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.13
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torch-geometric==2.6.1
torchaudio==2.5.0
torchdata==0.9.0
torchtext==0.18.0
torchvision==0.20.1
tornado==6.5.1
tqdm==4.66.5
transformers==4.49.0
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.67
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.37.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.6.0
yacs==0.1.8
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0
